+++ vck/src/commonMain/kotlin/at/asitplus/wallet/lib/agent/Validator.kt
@@ -31,6 +31,7 @@ import at.asitplus.wallet.lib.cbor.VerifyCoseSignatureWithKeyFun
 import at.asitplus.wallet.lib.data.*
 import at.asitplus.wallet.lib.data.rfc.tokenStatusList.StatusListTokenPayload
 import at.asitplus.wallet.lib.data.rfc.tokenStatusList.primitives.TokenStatus
+import at.asitplus.wallet.lib.data.rfc.tokenStatusList.primitives.TokenStatusValidationResult
 import at.asitplus.wallet.lib.iso.*
 import at.asitplus.wallet.lib.jws.*
 import io.github.aakira.napier.Napier
@@ -84,7 +85,7 @@ class Validator(
     private val resolveStatusListToken: StatusListTokenResolver? = null,
     /**
      * The function [tokenStatusResolver] should check the status mechanisms in a given status claim in order to
-     * evaluate the token status.
+     * extract the token status.
      */
     private val tokenStatusResolver: TokenStatusResolver = resolveStatusListToken?.toTokenStatusResolver(
         verifyJwsObjectIntegrity = verifyJwsObject,
@@ -94,6 +95,10 @@ class Validator(
     ) ?: TokenStatusResolver {
         KmmResult.success(TokenStatus.Valid)
     },
+    private val acceptedTokenStatuses: Set<TokenStatus> = setOf(TokenStatus.Valid),
+    private val tokenStatusValidator: TokenStatusValidator = tokenStatusResolver.toTokenStatusValidator(
+        acceptedTokenStatuses
+    ),
     private val vcJwsTimelinessValidator: VcJwsTimelinessValidator = VcJwsTimelinessValidator(
         timeLeeway = timeLeeway,
         clock = clock,
@@ -125,10 +130,10 @@ class Validator(
     /**
      * Checks the revocation state of the passed credential.
      */
-    suspend fun checkRevocationStatus(storeEntry: SubjectCredentialStore.StoreEntry) = tokenStatusResolver(storeEntry)
-    suspend fun checkRevocationStatus(issuerSigned: IssuerSigned): KmmResult<TokenStatus>? = tokenStatusResolver(issuerSigned)
-    suspend fun checkRevocationStatus(sdJwt: VerifiableCredentialSdJwt): KmmResult<TokenStatus>? = tokenStatusResolver(sdJwt)
-    suspend fun checkRevocationStatus(vcJws: VerifiableCredentialJws): KmmResult<TokenStatus>?  = tokenStatusResolver(vcJws)
+    suspend fun checkRevocationStatus(storeEntry: SubjectCredentialStore.StoreEntry): TokenStatusValidationResult = tokenStatusValidator(storeEntry)
+    suspend fun checkRevocationStatus(issuerSigned: IssuerSigned): TokenStatusValidationResult = tokenStatusValidator(issuerSigned)
+    suspend fun checkRevocationStatus(sdJwt: VerifiableCredentialSdJwt): TokenStatusValidationResult = tokenStatusValidator(sdJwt)
+    suspend fun checkRevocationStatus(vcJws: VerifiableCredentialJws): TokenStatusValidationResult = tokenStatusValidator(vcJws)
 
     /**
      * Validates the content of a JWS, expected to contain a Verifiable Presentation.
@@ -168,20 +173,11 @@ class Validator(
         }.map {
             VcJwsVerificationResultWrapper(
                 vcJws = it,
-                tokenStatus = tokenStatusResolver(
-                    CredentialWrapper.VcJws(it)
-                ),
+                tokenStatus = checkRevocationStatus(it),
                 timelinessValidationSummary = credentialTimelinessValidator(it),
             )
         }.groupBy {
-            it.timelinessValidationSummary.isSuccess && it.tokenStatus?.let {
-                // The library should probably not implicitly consider credentials valid if it isn't clear.
-                // Valid credentials should probably require no further considerations.
-                // Only consider TokenStatus.Valid is therefore implicitly considered valid.
-                //  - If other credentials shall be accepted, those can be selected from `untimelyVerifiableCredentials`
-                //      - selection should consider their token and expiration time status.
-                it.getOrNull() == TokenStatus.Valid
-            } ?: true
+            it.timelinessValidationSummary.isSuccess && it.tokenStatus is TokenStatusValidationResult.Valid
         }
 
         val vp = VerifiablePresentationParsed(
@@ -405,7 +401,7 @@ class Validator(
             !validationSummary.contentSemanticsValidationSummary.isSuccess -> InvalidStructure(input)
             validationSummary.subjectMatchingResult?.isSuccess == false -> ValidationError(input)
             validationSummary.isSuccess -> SuccessJwt(validationSummary.payload)
-            else -> ValidationError(input) // this branch shouldn't be executed happen anyway
+            else -> ValidationError(input) // this branch shouldn't be executed anyway
         }
     }