+++ vck/src/commonMain/kotlin/at/asitplus/wallet/lib/data/VcJwsVerificationResultWrapper.kt
@@ -1,11 +1,10 @@
 package at.asitplus.wallet.lib.data
 
-import at.asitplus.KmmResult
 import at.asitplus.wallet.lib.agent.validation.CredentialTimelinessValidationSummary
-import at.asitplus.wallet.lib.data.rfc.tokenStatusList.primitives.TokenStatus
+import at.asitplus.wallet.lib.data.rfc.tokenStatusList.primitives.TokenStatusValidationResult
 
 data class VcJwsVerificationResultWrapper(
     val vcJws: VerifiableCredentialJws,
-    val tokenStatus: KmmResult<TokenStatus>?,
+    val tokenStatus: TokenStatusValidationResult,
     val timelinessValidationSummary: CredentialTimelinessValidationSummary.VcJws,
 )