+++ vck/src/commonMain/kotlin/at/asitplus/wallet/lib/data/rfc/tokenStatusList/primitives/TokenStatusValidationResult.kt
@@ -1,0 +1,9 @@
+package at.asitplus.wallet.lib.data.rfc.tokenStatusList.primitives
+
+sealed interface TokenStatusValidationResult {
+    data class Rejected(val throwable: Throwable) : TokenStatusValidationResult
+
+    data class Invalid(val tokenStatus: TokenStatus) : TokenStatusValidationResult
+
+    data class Valid(val tokenStatus: TokenStatus?) : TokenStatusValidationResult
+}